+++ docs/source/CHANGELOG.rst
@@ -2,13 +2,25 @@
 Changelog
 =========
 
-Unreleased
-----------
+
+1.9.0 - 2021-12-22
+------------------
 
 Added
 ^^^^^
 * New endpoint for Convert:
-   * ``GET /sapi/v1/convert/tradeFlow`` to support user query convert trade history records
+  * ``GET /sapi/v1/convert/tradeFlow`` to support user query convert trade history records* New endpoint for Rebate:
+* New endpoint for Rebate:
+  * ``GET /sapi/v1/rebate/taxQuery`` to support user query spot rebate history records
+* New endpoints for Margin:
+  * ``GET /sapi/v1/margin/crossMarginData`` to get cross margin fee data collection
+  * ``GET /sapi/v1/margin/isolatedMarginData`` to get isolated margin fee data collection
+  * ``GET /sapi/v1/margin/isolatedMarginTier`` to get isolated margin tier data collection
+* New endpoints for NFT:
+  * ``GET /sapi/v1/nft/history/transactions`` to get NFT transaction history
+  * ``GET /sapi/v1/nft/history/deposit`` to get NFT deposit history
+  * ``GET /sapi/v1/nft/history/withdraw`` to get NFT withdraw history
+  * ``GET /sapi/v1/nft/user/getAsset`` to get NFT asset
 
 1.8.0 - 2021-11-25
 ------------------